test: address markdown drag review feedback
